Move opentrons_modules fields to ModuleContainers object
