+++ emulation_system/tests/e2e/docker_interface/module_containers.py
@@ -5,6 +5,10 @@ from typing import List, Optional, Set
 
 from docker.models.containers import Container  # type: ignore[import]
 
+from emulation_system.compose_file_creator.config_file_settings import RepoToBuildArgMapping
+from tests.e2e.test_definition.build_arg_configurations import BuildArgConfigurations
+from tests.e2e.utilities.helper_functions import get_mounts
+
 
 @dataclass
 class ModuleContainers:
@@ -16,6 +20,39 @@ class ModuleContainers:
     firmware_emulation_heater_shaker_modules: List[Container]
     firmware_emulation_magnetic_modules: List[Container]
     firmware_emulation_temperature_modules: List[Container]
+
+    modules_builder: Optional[Container]
+
+    @property
+    def opentrons_modules_builder_created(self) -> bool:
+        """Whether or not the opentrons-modules builder container was created."""
+        return self.modules_builder is not None
+
+    @property
+    def local_opentrons_modules_mounted(self) -> bool:
+        """Whether or not the opentrons-modules builder container has local source mounted."""
+        if self.modules_builder is None:
+            return False
+
+        modules_builder_mounts = get_mounts(self.modules_builder)
+        return (
+            self.opentrons_modules_builder_created
+            and modules_builder_mounts is not None
+            and any(
+                mount["Destination"] == "/opentrons-modules"
+                for mount in modules_builder_mounts
+            )
+        )
+
+    @property
+    def opentrons_modules_build_args(self) -> "BuildArgConfigurations":
+        """Returns BuildArgConfigurations object representing where source was pulled from."""
+        return BuildArgConfigurations.parse_build_args(
+            self.modules_builder,
+            "opentrons-modules/archive/refs/heads/edge.zip",
+            RepoToBuildArgMapping.OPENTRONS_MODULES,
+        )
+
 
     @property
     def hardware_emulation_thermocycler_module_names(self) -> Set[str]:
