+++ emulation_system/tests/e2e/docker_interface/ot3_containers.py
@@ -31,7 +31,7 @@ class OT3SystemUnderTest:
     # but want to check that with an assert and not a type error
     firmware_builder: Optional[Container]
     monorepo_builder: Optional[Container]
-    modules_builder: Optional[Container]
+
 
     @property
     def monorepo_builder_created(self) -> bool:
@@ -93,35 +93,6 @@ class OT3SystemUnderTest:
             RepoToBuildArgMapping.OT3_FIRMWARE,
         )
 
-    @property
-    def opentrons_modules_builder_created(self) -> bool:
-        """Whether or not the opentrons-modules builder container was created."""
-        return self.modules_builder is not None
-
-    @property
-    def local_opentrons_modules_mounted(self) -> bool:
-        """Whether or not the opentrons-modules builder container has local source mounted."""
-        if self.modules_builder is None:
-            return False
-
-        modules_builder_mounts = get_mounts(self.modules_builder)
-        return (
-            self.opentrons_modules_builder_created
-            and modules_builder_mounts is not None
-            and any(
-                mount["Destination"] == "/opentrons-modules"
-                for mount in modules_builder_mounts
-            )
-        )
-
-    @property
-    def opentrons_modules_build_args(self) -> "BuildArgConfigurations":
-        """Returns BuildArgConfigurations object representing where source was pulled from."""
-        return BuildArgConfigurations.parse_build_args(
-            self.modules_builder,
-            "opentrons-modules/archive/refs/heads/edge.zip",
-            RepoToBuildArgMapping.OPENTRONS_MODULES,
-        )
 
     @property
     def containers_with_entrypoint_script(self) -> List[Container]:
